onvif_media_snapshot.c - changed size type from double to int

+++ src/onvif_media_snapshot.c
@@ -26,7 +26,7 @@ void OnvifSnapshot__destroy(OnvifSnapshot * self){
     }
 }
 
-double OnvifSnapshot__get_size(OnvifSnapshot * self){
+int OnvifSnapshot__get_size(OnvifSnapshot * self){
     return self->size;
 }
